More migrating of compress/decompression code into its own files to help hide it away more and cleanup its tentacles
