+++ 3rdParty/LightPcapNg/LightPcapNg/include/light_file.h
@@ -1,0 +1,42 @@
+// light_file.h
+// Created on: Aug 13, 2019
+
+// Copyright (c) 2019 TMEIC Corporation - Robert Kriener
+
+// Permission is hereby granted, free of charge, to any person obtaining a copy
+// of this software and associated documentation files (the "Software"), to deal
+// in the Software without restriction, including without limitation the rights
+// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+// copies of the Software, and to permit persons to whom the Software is
+// furnished to do so, subject to the following conditions:
+
+// The above copyright notice and this permission notice shall be included in all
+// copies or substantial portions of the Software.
+
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+// SOFTWARE.
+
+#ifndef INCLUDE_LIGHT_FILE_H_
+#define INCLUDE_LIGHT_FILE_H_
+
+#include <stdio.h>
+#include "light_compression.h"
+
+typedef _compression_t *light_compression;
+typedef _decompression_t *light_decompression;
+
+typedef struct
+{
+	FILE* file;
+	light_compression compression_context;
+	light_decompression decompression_context;
+
+}
+light_file_t, *light_file;
+
+#endif /* INCLUDE_LIGHT_FILE_H_ */
