+++ 3rdParty/LightPcapNg/LightPcapNg/include/light_internal.h
@@ -30,60 +30,6 @@
 #include <stdint.h>
 #include <assert.h>
 
-#if defined(USE_Z_STD)
-#include <zstd.h>      // presumes zstd library is installed
-
-//An ethernet packet should only ever be up to 1500 bytes + some header crap
-//We also expect some ovehead for the pcapng blocks which contain the ethernet packets
-//so allocate 1700 bytes as the max input size we expect in a single shot
-#define COMPRESSION_BUFFER_IN_MAX_SIZE 1700
-
-//This is the z-std compression type I would call it z-std type and realias 
-//2x but complier won't let me do that across bounds it seems
-//So I gave it a generic "light" name....
-struct light_compression_t
-{
-	uint32_t* buffer_in;
-	uint32_t* buffer_out;
-	size_t buffer_in_max_size;
-	size_t buffer_out_max_size;
-	int compression_level;
-	ZSTD_CCtx* cctx;
-};
-
-struct light_decompression_t
-{
-	uint32_t* buffer_in;
-	uint32_t* buffer_out;
-	size_t buffer_in_max_size;
-	size_t buffer_out_max_size;
-	ZSTD_DCtx* dctx;
-	int outputReady;
-	ZSTD_outBuffer output;
-	ZSTD_inBuffer input;
-};
-
-//I really would like to use _compression_t everywhere and realias that to light_compression * in lightpcapng.h
-//but the complier does not seem to allow me to use 2x alias across files
-typedef struct light_compression_t _compression_t;
-typedef struct light_decompression_t _decompression_t;
-
-//Setup some other compression
-#elif defined(USE_THIS_COMPRESSION_INSTEAD)
-
-//No compression
-#else
-
-typedef void _compression_t;
-
-#endif
-
-void light_free_compression_context(_compression_t* context);
-_compression_t * light_get_compression_context(int compression_level);
-
-void light_free_decompression_context(_decompression_t* context);
-_decompression_t * light_get_decompression_context();
-
 struct _light_pcapng {
 	uint32_t block_type;
 	uint32_t block_total_length;
