+++ 3rdParty/LightPcapNg/LightPcapNg/src/light_platform.c
@@ -23,10 +23,13 @@
 
 #include "light_platform.h"
 #include "light_internal.h"
+#include "light_compression.h"
 
 #include <stdlib.h>
 #include <stdio.h>
 #include <string.h>
+
+
 
 #ifdef UNIVERSAL
 
@@ -63,7 +66,7 @@ light_file light_open(const char *file_name, const __read_mode_t mode)
 	switch (mode) {
 	case LIGHT_OREAD:
 	{
-		if (strstr(file_name, ".zstd"))
+		if (light_is_compressed_file(file_name))
 		{
 			return light_open_decompression(file_name, mode);
 		}
@@ -130,76 +133,10 @@ size_t light_read(light_file fd, void *buf, size_t count)
 		size_t bytes_read = fread(buf, 1, count, fd->file);
 		return  bytes_read != count ? -1 : bytes_read;
 	}
-#if defined(USE_Z_STD)
+	else
 	{
-		//Decompression is a little more complex
-		//Need to manage reading bytes from orignal file
-		//Decompressing those into a buffer
-		//Then reading the selected number of bytes from the buffer
-		//Once whole buffer is consumed we need to read and decompress next chunk from file
-
-		size_t bytes_read = 0;
-
-		while (bytes_read < count)
-		{
-			if (fd->decompression_context->outputReady == 0)
-			{
-				//Check if we need to grab a new chunk from the actual file
-				//If we read all the input then yes, we need to do that
-				if (fd->decompression_context->input.pos >= fd->decompression_context->input.size)
-				{
-					//Read a decompress a chunk
-					size_t bytes_read_file = fread(fd->decompression_context->buffer_in, 1, fd->decompression_context->buffer_in_max_size, fd->file);
-					if (bytes_read_file < fd->decompression_context->buffer_in_max_size && bytes_read_file == 0 && feof(fd->file))
-						return EOF;
-					fd->decompression_context->input.src = fd->decompression_context->buffer_in;
-					fd->decompression_context->input.size = bytes_read_file;
-					fd->decompression_context->input.pos = 0;
-				}
-				//Decompress into the output buffer and use this buffer to actually get our results
-				fd->decompression_context->output.dst = fd->decompression_context->buffer_out;
-				fd->decompression_context->output.size = fd->decompression_context->buffer_out_max_size;
-				fd->decompression_context->output.pos = 0;
-
-				size_t const remaining = ZSTD_decompressStream(fd->decompression_context->dctx, &fd->decompression_context->output, &fd->decompression_context->input);
-				assert(!ZSTD_isError(remaining));
-
-				//Re-use the output class to track our own consumption
-				fd->decompression_context->output.size = fd->decompression_context->output.pos;
-				fd->decompression_context->output.pos = 0;
-				fd->decompression_context->outputReady = 1;
-			}
-
-			//Now get bytes from our output buffer
-
-			//Case we need everything or less than that which was decoded
-			int needToRead = count - bytes_read;
-			int remaining = (fd->decompression_context->output.size - fd->decompression_context->output.pos);
-			if (needToRead <= remaining)
-			{
-				memcpy((uint8_t*)buf + bytes_read, (uint8_t*)fd->decompression_context->output.dst + fd->decompression_context->output.pos, needToRead);
-				fd->decompression_context->output.pos += needToRead;
-				bytes_read += needToRead;
-			}
-
-			//Case we need more than that which was decoded
-			if (needToRead > remaining)
-			{
-				memcpy((uint8_t*)buf + bytes_read, (uint8_t*)fd->decompression_context->output.dst + fd->decompression_context->output.pos, remaining);
-				bytes_read += remaining;
-				fd->decompression_context->output.pos += remaining;
-			}
-
-			//We have consumed everything - set next call to decompress a new chunk
-			if (fd->decompression_context->output.pos == fd->decompression_context->output.size)
-				fd->decompression_context->outputReady = 0;
-
-		}
-
-
-		return bytes_read;
+		return light_read_compressed(fd, buf, count);
 	}
-#endif 
 }
 
 size_t light_write(light_file fd, const void *buf, size_t count)
@@ -209,33 +146,10 @@ size_t light_write(light_file fd, const void *buf, size_t count)
 		size_t bytes_written = fwrite(buf, 1, count, fd->file);
 		return  bytes_written != count ? -1 : bytes_written;
 	}
-	#if defined(USE_Z_STD)
+	else
 	{
-		//Do compression here!
-		/* Set the input buffer to what we just read.
-		* We compress until the input buffer is empty, each time flushing the
-		* output.
-		*/
-		ZSTD_inBuffer input = { buf, count, 0 };
-		int finished;
-		do
-		{
-			/* Compress into the output buffer and write all of the output to
-			* the file so we can reuse the buffer next iteration.
-			*/
-			ZSTD_outBuffer output = { fd->compression_context->buffer_out, fd->compression_context->buffer_out_max_size, 0 };
-			size_t const remaining = ZSTD_compressStream2(fd->compression_context->cctx, &output, &input, ZSTD_e_continue);
-			assert(!ZSTD_isError(remaining));
-			fwrite(output.dst, 1, output.pos, fd->file);
-			/* If we're on the last chunk we're finished when zstd returns 0,
-			 * We're finished when we've consumed all the input.
-			 */
-			finished = (input.pos == input.size);
-		} while (!finished);
-
-		return count;
+		return light_write_compressed(fd, buf, count);
 	}
-	#endif 
 }
 
 size_t light_size(light_file fd)
@@ -252,24 +166,7 @@ size_t light_size(light_file fd)
 
 int light_close(light_file fd)
 {
-	//Wrap up the compression here
-	if (fd->compression_context)
-	{
-	#if defined(USE_Z_STD)
-		ZSTD_inBuffer input = { 0,0,0 };
-
-		int remaining = 1;
-
-		while (remaining != 0)
-		{
-			ZSTD_outBuffer output = { fd->compression_context->buffer_out, fd->compression_context->buffer_out_max_size, 0 };
-			remaining = ZSTD_compressStream2(fd->compression_context->cctx, &output, &input, ZSTD_e_end);
-			fwrite(output.dst, 1, output.pos, fd->file);
-		}
-	#endif
-	}
-	light_free_compression_context(fd->compression_context);
-
+	light_close_compresssed(fd);
 	return fclose(fd->file);
 }
